+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/NestingConfiguration.swift
@@ -13,6 +13,8 @@ struct NestingConfiguration: RuleConfiguration {
     private(set) var checkNestingInClosuresAndStatements = true
     @ConfigurationElement(key: "always_allow_one_type_in_functions")
     private(set) var alwaysAllowOneTypeInFunctions = false
+    @ConfigurationElement(key: "ignore_nesting_typealias_and_associatedtype")
+    private(set) var ignoreNestingTypealiasAndAssociatedtype = false
 
     func severity(with config: Severity, for level: Int) -> ViolationSeverity? {
         if let error = config.error, level > error {