+++ Tests/SwiftLintFrameworkTests/NestingRuleTests.swift
@@ -368,4 +368,45 @@ class NestingRuleTests: SwiftLintTestCase {
 
         verifyRule(description, ruleConfiguration: ["check_nesting_in_closures_and_statements": false])
     }
+    
+    func testNestingWithoutTypealiasAndAssociatedtype() {
+        var nonTriggeringExamples = NestingRule.description.nonTriggeringExamples
+        nonTriggeringExamples.append(contentsOf: ["class", "struct", "enum"].flatMap { type -> [Example] in
+            [
+                .init("""
+                    \(type) Example_0 {
+                        \(type) Example_1 {
+                            typealias Example_2_Type = Example_2.Type
+                        }
+                        \(type) Example_2 {
+                           
+                        }
+                    }
+                """),
+            ]
+        })
+        
+        let triggeringExamples = ["class", "struct", "enum"].flatMap { type -> [Example] in
+            [
+                .init("""
+                    \(type) Example_0 {
+                        \(type) Example_1 {
+                            \(type) Example_2 {
+                               
+                            }
+                        }
+                    }
+                """),
+            ]
+        }
+        
+        let descripton = RuleDescription(identifier: NestingRule.description.identifier,
+                                         name: NestingRule.description.name,
+                                         description: NestingRule.description.description,
+                                         kind: .metrics,
+                                         nonTriggeringExamples: nonTriggeringExamples,
+                                         triggeringExamples: triggeringExamples)
+        
+        verifyRule(descripton, ruleConfiguration: ["ignore_nesting_typealias_and_associatedtype": true])
+    }
 }